llm-prompt-engine-api/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.32.0
prometheus-fastapi-instrumentator==7.0.0
python-json-logger==2.0.7
pyyaml==6.0.1
openapi-spec-validator==0.6.1